Added a #DEFINE for inverted relay output
I am using a relay driven with an optoisolator and transistor.  This
means when I want to enable the relay, I set the output to LOW.  This is
the opposite of driving the relay just with the digital pin alone.

If you are just driving a plain relay then comment out the define
statement for RELAY_INVERTED